pyproject.toml
[build-system]
requires = ["poetry-core=1.7"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "tg_bot"
version = "0.0.1"
description = "tg_bot"
authors = ["Konstantin Klepikov <oformleno@gmail.com>"]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
passlib = "1.7.4"
toml = "0.10.2"
fastapi = "^0.104.1"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
pydantic = "^2.4.2"
python-multipart = ">=0.0.6,<0.1.0"
pydantic-settings = "^2.5.2"
pydantic-core = "^2.23.4"
annotated-types = "^0.7.0"
pyyaml = "^6.0.2"
